Fixed GDI handles leak - it was the file icons not destroyed
The problem: Icon.FromHandle doesn't take ownership of the icon handle (shfi.hIcon). When I call this method, it creates a new Icon object that references the handle, but it doesn't destroy the handle when the Icon is garbage collected. This is causing your GDI handle leak.
Solution  - created a clone of icon to managed space, and destroyed the unmanaged file icon from heap, by calling win api to destroy it